Restructure vader/test_helpers, fix types in test
