+++ tests/custom_highlights_spec.lua
@@ -3,8 +3,10 @@ local compat = require("winmove.compat")
 local config = require("winmove.config")
 local highlight = require("winmove.highlight")
 local vader = require("winmove.util.vader")
+local test_helpers = require("winmove.util.test_helpers")
 
 local given = vader.given
+local make_layout = test_helpers.make_layout
 
 describe("custom highlights", function()
     if not compat.has("nvim-0.9.0") then
@@ -22,7 +24,7 @@ describe("custom highlights", function()
         })
 
         given("", function()
-            vader.make_layout({
+            make_layout({
                 "row",
                 { "leaf", "leaf" },
             })
@@ -64,7 +66,7 @@ describe("custom highlights", function()
         })
 
         given("", function()
-            vader.make_layout({
+            make_layout({
                 "row",
                 { "leaf", "leaf" },
             })
