+++ tests/show_help_spec.lua
@@ -2,8 +2,10 @@ local winmove = require("winmove")
 local config = require("winmove.config")
 local float = require("winmove.float")
 local vader = require("winmove.util.vader")
+local test_helpers = require("winmove.util.test_helpers")
 
 local given = vader.given
+local make_layout = test_helpers.make_layout
 
 describe("shows floating window help", function()
     -- Ensure default keymaps
@@ -16,7 +18,7 @@ describe("shows floating window help", function()
 
     it("shows help in resize mode", function()
         given("", function()
-            vader.make_layout({
+            make_layout({
                 "row",
                 { "leaf", "leaf" },
             })
